Refactor authentication hooks and remove deprecated loading screen
- Updated import paths for authentication hooks from `use-auth-v2` to `use-auth` across multiple components.
- Removed the `loading-screen.tsx` component as it is no longer needed.
- Replaced the old `use-auth` hook implementation with a new version that utilizes React Query for better state management and API interactions.
- Deleted the mock data and related functionality from `useDefects.ts` to streamline the codebase.
