+++ src/App.tsx
@@ -5,10 +5,8 @@ import OperatorPage from "@/pages/operator-page";
 import ProtectedRoute from "@/components/protected-route";
 import AuthGuard from "@/components/auth-guard";
 import QCPage from "@/pages/qc-page";
-import { ManagerDashboard } from "./components/ManagerDashboard";
-import { OperatorDashboard } from "./components/OperatorDashboard";
 
-import { useAuth } from "./hooks/auth/use-auth-v2";
+import { useAuth } from "./hooks/auth/use-auth";
 import { ROLES } from "./contants/auth";
 
 import type { RoleType } from "./types/auth";
@@ -20,13 +18,13 @@ function App() {
     switch (role) {
       case ROLES.OPERATOR:
       case ROLES.VIEWER:
-        return <OperatorPage />; //V2
+        return <OperatorPage />;
       case ROLES.INSPECTOR:
-        return <QCPage />; // V2
+        return <QCPage />;
       case ROLES.SUPERADMIN:
-        return <ManagerDashboard />; // V1
+        return <>TODO</>;
       default:
-        return <OperatorDashboard />; // V1
+        return <>TODO</>;
     }
   };
