--- src/components/ConfirmationDialog.tsx
@@ -1,123 +1,0 @@
-import React from 'react';
-import { X, AlertCircle } from 'lucide-react';
-
-interface ConfirmationDialogProps {
-  isOpen: boolean;
-  title: string;
-  message: string;
-  confirmText?: string;
-  cancelText?: string;
-  onConfirm: () => void;
-  onCancel: () => void;
-  type?: 'info' | 'warning' | 'success';
-  icon?: React.ReactNode;
-}
-
-export const ConfirmationDialog: React.FC<ConfirmationDialogProps> = ({
-  isOpen,
-  title,
-  message,
-  confirmText = 'ยืนยัน',
-  cancelText = 'ยกเลิก',
-  onConfirm,
-  onCancel,
-  type = 'info',
-  icon
-}) => {
-  if (!isOpen) return null;
-
-  const getTypeStyles = () => {
-    switch (type) {
-      case 'warning':
-        return {
-          iconBg: 'bg-orange-100',
-          iconColor: 'text-orange-600',
-          confirmBtn: 'bg-orange-600 hover:bg-orange-700'
-        };
-      case 'success':
-        return {
-          iconBg: 'bg-green-100',
-          iconColor: 'text-green-600',
-          confirmBtn: 'bg-green-600 hover:bg-green-700'
-        };
-      default:
-        return {
-          iconBg: 'bg-blue-100',
-          iconColor: 'text-blue-600',
-          confirmBtn: 'bg-blue-600 hover:bg-blue-700'
-        };
-    }
-  };
-
-  const styles = getTypeStyles();
-
-  const handleBackdropClick = (e: React.MouseEvent) => {
-    if (e.target === e.currentTarget) {
-      onCancel();
-    }
-  };
-
-  const handleKeyDown = (e: React.KeyboardEvent) => {
-    if (e.key === 'Escape') {
-      onCancel();
-    } else if (e.key === 'Enter') {
-      onConfirm();
-    }
-  };
-
-  return (
-    <div 
-      className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center p-4 z-50"
-      onClick={handleBackdropClick}
-    >
-      <div className="bg-white rounded-xl shadow-2xl max-w-md w-full animate-in fade-in-0 zoom-in-95 duration-200">
-        {/* Header */}
-        <div className="p-6 border-b border-gray-200">
-          <div className="flex justify-between items-center">
-            <h3 className="text-lg font-semibold text-gray-900">{title}</h3>
-            <button
-              onClick={onCancel}
-              className="p-2 hover:bg-gray-100 rounded-full transition-colors"
-              aria-label="ปิด"
-            >
-              <X className="w-5 h-5 text-gray-500" />
-            </button>
-          </div>
-        </div>
-
-        {/* Content */}
-        <div className="p-6">
-          <div className="flex items-start space-x-4">
-            {/* Icon */}
-            <div className={`flex-shrink-0 w-12 h-12 ${styles.iconBg} rounded-full flex items-center justify-center`}>
-              {icon || <AlertCircle className={`w-6 h-6 ${styles.iconColor}`} />}
-            </div>
-            
-            {/* Message */}
-            <div className="flex-1">
-              <p className="text-gray-700 leading-relaxed">{message}</p>
-            </div>
-          </div>
-        </div>
-
-        {/* Footer */}
-        <div className="p-6 border-t border-gray-200 flex space-x-3">
-          <button
-            onClick={onCancel}
-            className="flex-1 bg-gray-200 hover:bg-gray-300 text-gray-800 py-2 px-4 rounded-lg font-medium transition-colors focus:ring-2 focus:ring-gray-500 focus:ring-offset-2"
-          >
-            {cancelText}
-          </button>
-          <button
-            onClick={onConfirm}
-            className={`flex-1 ${styles.confirmBtn} text-white py-2 px-4 rounded-lg font-medium transition-colors focus:ring-2 focus:ring-offset-2`}
-            onKeyDown={handleKeyDown}
-            autoFocus
-          >
-            {confirmText}
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-};
