--- src/components/ManagerDashboard.tsx
@@ -1,202 +1,0 @@
-import React, { useState } from 'react';
-import { Layout } from './Layout';
-import { useDefects } from '../hooks/useDefects';
-import { BarChart3, TrendingUp, Download, Calendar, Filter } from 'lucide-react';
-import { ReportModal } from './ReportModal';
-
-export const ManagerDashboard: React.FC = () => {
-  const { defects } = useDefects();
-  const [showReportModal, setShowReportModal] = useState(false);
-
-  const totalDefects = defects.length;
-  const pendingDefects = defects.filter(d => d.state === 'pending' || d.state === 'acknowledged').length;
-  const approvedDefects = defects.filter(d => d.state === 'approved').length;
-  const rejectedDefects = defects.filter(d => d.state === 'rejected').length;
-
-  const defectsByType = defects.reduce((acc, defect) => {
-    acc[defect.status] = (acc[defect.status] || 0) + 1;
-    return acc;
-  }, {} as Record<string, number>);
-
-  const defectsByLine = defects.reduce((acc, defect) => {
-    const line = `Line ${defect.productionLine}`;
-    acc[line] = (acc[line] || 0) + 1;
-    return acc;
-  }, {} as Record<string, number>);
-
-  return (
-    <Layout title="Manager Dashboard">
-      <div className="space-y-6">
-        {/* Header Actions */}
-        <div className="flex justify-between items-center">
-          <h2 className="text-2xl font-bold text-gray-900">Production Quality Overview</h2>
-          <button
-            onClick={() => setShowReportModal(true)}
-            className="bg-blue-600 hover:bg-blue-700 text-white px-4 py-2 rounded-lg font-medium flex items-center gap-2 transition-colors"
-          >
-            <Download className="w-4 h-4" />
-            Generate Report
-          </button>
-        </div>
-
-        {/* Key Metrics */}
-        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-6">
-          <div className="bg-white rounded-lg border border-gray-200 p-6">
-            <div className="flex items-center justify-between">
-              <div>
-                <p className="text-sm font-medium text-gray-600">Total Defects</p>
-                <p className="text-3xl font-bold text-gray-900">{totalDefects}</p>
-              </div>
-              <div className="w-12 h-12 bg-blue-100 rounded-lg flex items-center justify-center">
-                <BarChart3 className="w-6 h-6 text-blue-600" />
-              </div>
-            </div>
-          </div>
-
-          <div className="bg-white rounded-lg border border-gray-200 p-6">
-            <div className="flex items-center justify-between">
-              <div>
-                <p className="text-sm font-medium text-gray-600">Pending Review</p>
-                <p className="text-3xl font-bold text-orange-600">{pendingDefects}</p>
-              </div>
-              <div className="w-12 h-12 bg-orange-100 rounded-lg flex items-center justify-center">
-                <Calendar className="w-6 h-6 text-orange-600" />
-              </div>
-            </div>
-          </div>
-
-          <div className="bg-white rounded-lg border border-gray-200 p-6">
-            <div className="flex items-center justify-between">
-              <div>
-                <p className="text-sm font-medium text-gray-600">Approved</p>
-                <p className="text-3xl font-bold text-green-600">{approvedDefects}</p>
-              </div>
-              <div className="w-12 h-12 bg-green-100 rounded-lg flex items-center justify-center">
-                <TrendingUp className="w-6 h-6 text-green-600" />
-              </div>
-            </div>
-          </div>
-
-          <div className="bg-white rounded-lg border border-gray-200 p-6">
-            <div className="flex items-center justify-between">
-              <div>
-                <p className="text-sm font-medium text-gray-600">Rejected</p>
-                <p className="text-3xl font-bold text-red-600">{rejectedDefects}</p>
-              </div>
-              <div className="w-12 h-12 bg-red-100 rounded-lg flex items-center justify-center">
-                <Filter className="w-6 h-6 text-red-600" />
-              </div>
-            </div>
-          </div>
-        </div>
-
-        {/* Charts and Analysis */}
-        <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-          {/* Defects by Type */}
-          <div className="bg-white rounded-lg border border-gray-200 p-6">
-            <h3 className="text-lg font-semibold text-gray-900 mb-4">Defects by Type</h3>
-            <div className="space-y-4">
-              {Object.entries(defectsByType).map(([type, count]) => (
-                <div key={type} className="flex items-center justify-between">
-                  <span className="text-sm text-gray-600">{type}</span>
-                  <div className="flex items-center space-x-2">
-                    <div className="w-32 bg-gray-200 rounded-full h-2">
-                      <div
-                        className="bg-blue-600 h-2 rounded-full"
-                        style={{ width: `${(count / totalDefects) * 100}%` }}
-                      ></div>
-                    </div>
-                    <span className="text-sm font-medium text-gray-900">{count}</span>
-                  </div>
-                </div>
-              ))}
-            </div>
-          </div>
-
-          {/* Defects by Production Line */}
-          <div className="bg-white rounded-lg border border-gray-200 p-6">
-            <h3 className="text-lg font-semibold text-gray-900 mb-4">Defects by Production Line</h3>
-            <div className="space-y-4">
-              {Object.entries(defectsByLine).map(([line, count]) => (
-                <div key={line} className="flex items-center justify-between">
-                  <span className="text-sm text-gray-600">{line}</span>
-                  <div className="flex items-center space-x-2">
-                    <div className="w-32 bg-gray-200 rounded-full h-2">
-                      <div
-                        className="bg-green-600 h-2 rounded-full"
-                        style={{ width: `${(count / totalDefects) * 100}%` }}
-                      ></div>
-                    </div>
-                    <span className="text-sm font-medium text-gray-900">{count}</span>
-                  </div>
-                </div>
-              ))}
-            </div>
-          </div>
-        </div>
-
-        {/* Recent Activity */}
-        <div className="bg-white rounded-lg border border-gray-200">
-          <div className="p-6 border-b border-gray-200">
-            <h3 className="text-lg font-semibold text-gray-900">Recent Defect Activity</h3>
-          </div>
-          <div className="overflow-x-auto">
-            <table className="min-w-full">
-              <thead className="bg-gray-50">
-                <tr>
-                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Production Line</th>
-                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Station</th>
-                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Product Code</th>
-                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Job Order Number</th>
-                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Status</th>
-                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">State</th>
-                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Timestamp</th>
-                </tr>
-              </thead>
-              <tbody className="bg-white divide-y divide-gray-200">
-                {defects.slice(0, 10).map((defect) => (
-                  <tr key={defect.id} className="hover:bg-gray-50">
-                    <td className="px-6 py-4 whitespace-nowrap text-sm font-medium text-gray-900">
-                      Line {defect.productionLine}
-                    </td>
-                    <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-600">{defect.station}</td>
-                    <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-600">{defect.productCode}</td>
-                    <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-600">{defect.jobOrderNumber}</td>
-                    <td className="px-6 py-4 whitespace-nowrap">
-                      <span className={`inline-flex px-2 py-1 text-xs font-medium rounded-full ${
-                        defect.status.includes('Defect') ? 'bg-orange-100 text-orange-800' :
-                        defect.status === 'Scrap' ? 'bg-red-100 text-red-800' :
-                        'bg-green-100 text-green-800'
-                      }`}>
-                        {defect.status}
-                      </span>
-                    </td>
-                    <td className="px-6 py-4 whitespace-nowrap">
-                      <span className={`inline-flex px-2 py-1 text-xs font-medium rounded-full ${
-                        defect.state === 'pending' ? 'bg-yellow-100 text-yellow-800' :
-                        defect.state === 'acknowledged' ? 'bg-blue-100 text-blue-800' :
-                        defect.state === 'approved' ? 'bg-green-100 text-green-800' :
-                        'bg-red-100 text-red-800'
-                      }`}>
-                        {defect.state.charAt(0).toUpperCase() + defect.state.slice(1)}
-                      </span>
-                    </td>
-                    <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-600">{defect.timestamp}</td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </div>
-        </div>
-      </div>
-
-      {/* Report Modal */}
-      {showReportModal && (
-        <ReportModal
-          defects={defects}
-          onClose={() => setShowReportModal(false)}
-        />
-      )}
-    </Layout>
-  );
-};
