--- src/hooks/auth/use-auth.ts
@@ -1,15 +1,0 @@
-import { createContext, useContext } from "react";
-
-import type { AuthContextType } from "@/contexts/AuthContext";
-
-export const AuthContext = createContext<AuthContextType | undefined>(
-  undefined
-);
-
-export const useAuth = (): AuthContextType => {
-  const context = useContext(AuthContext);
-  if (!context) {
-    throw new Error("useAuth must be used within an AuthProvider");
-  }
-  return context;
-};
