+++ src/pages/login-page/login-form.tsx
@@ -12,7 +12,7 @@ import {
 } from "@/components/ui/form";
 
 import { loginSchema } from "./schema";
-import { useAuth } from "@/hooks/auth/use-auth-v2";
+import { useAuth } from "@/hooks/auth/use-auth";
 import { cn } from "@/lib/utils";
 
 import type { LoginFormType } from "./types";
