+++ src/pages/qc-page/components/review-history-table.tsx
@@ -11,7 +11,7 @@ import {
 
 import { useProductionLineOptions } from "@/hooks/option/use-production-line-option";
 import { HISTORY_COLUMNS } from "../constants/history-columns";
-import { useAuth } from "@/hooks/auth/use-auth-v2";
+import { useAuth } from "@/hooks/auth/use-auth";
 import { useDefectOptionAPI } from "@/hooks/option/use-defect-option";
 import { useReviewAPI } from "@/hooks/review/use-review";
 import { REVIEW_STATE_OPTION, REVIEW_STATE } from "@/contants/review";
