+++ src/pages/qc-page/components/review-table.tsx
@@ -10,7 +10,7 @@ import {
 } from "@/components/ui/select";
 
 import { REVIEW_COLUMNS } from "../constants/review-columns";
-import { useAuth } from "@/hooks/auth/use-auth-v2";
+import { useAuth } from "@/hooks/auth/use-auth";
 import { useProductionLineOptions } from "@/hooks/option/use-production-line-option";
 // import { ALL_OPTION } from "@/contants/option";
 // import { useDefectOptionAPI } from "@/hooks/option/use-defect-option";
