+++ src/pages/qc-page/components/waiting-review-table.tsx
@@ -11,7 +11,7 @@ import {
 
 import { ALL_OPTION } from "@/contants/option";
 import { REVIEW_STATE } from "@/contants/review";
-import { useAuth } from "@/hooks/auth/use-auth-v2";
+import { useAuth } from "@/hooks/auth/use-auth";
 import { useDefectOptionAPI } from "@/hooks/option/use-defect-option";
 import { useProductionLineOptions } from "@/hooks/option/use-production-line-option";
 import { useReviewAPI } from "@/hooks/review/use-review";
